backend/resume-optimizer/pyproject.toml
[project]
name = "resume-optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.66.0",
    "pymupdf>=1.27.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
]
